CrowFall-Map/requirements.txt
asgiref==3.5.2
attrs==22.1.0
beautifulsoup4==4.11.1
black==22.1.0
bs4==0.0.1
certifi==2021.10.8
chardet==4.0.0
click==8.0.3
dj-database-url==1.0.0
dj-email-url==1.0.5
Django==4.1
django-cache-url==3.4.2
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
environ==1.0
environs==9.3.2
flake8==4.0.1
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
marshmallow==3.14.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
phonenumbers==8.12.41
Pillow==9.2.0
platformdirs==2.4.1
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.2
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==7.1.2
python-dateutil==2.8.1
python-dotenv==0.19.2
pytz==2021.3
requests==2.25.1
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.0
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==6.2.0
